backend/requirements.txt
flask
flask-cors
flask-jwt-extended
flask-sqlalchemy
pymysql
pyodbc
python-dotenv
werkzeug
